+++ src/functions/arithmetic/decimal.js
@@ -12,6 +12,11 @@ class LimeFunctionDecimal extends LimeFunction {
 
     // Binary operation
     this.operations.b = [
+      'tf(.+)',
+      'tf(./)',
+      'tf(.*)',
+      'tf(.-)',
+
       'eb(int,int)',
     ];