+++ src/functions/arithmetic/divide.js
@@ -40,12 +40,14 @@ class LimeFunctionDivide extends LimeFunction {
         step.bs(this.lime.build('rational')(step.left, step.right));
       }
     });
+
     this.algorithms.set('b(rat,rat)', (step) => {
       step.bs(this.lime.build('rational')(
         this.lime.direct([step.left.nPlace, '*', step.right.dPlace]),
         this.lime.direct([step.left.dPlace, '*', step.right.nPlace]),
       ));
     });
+
     this.algorithms.set('b(comp,comp)', (step) => {
       step.bs(this.lime.build('complex')(
         this.lime.direct(['re', '(', step.left, '*', 'conj', '(', step.right, ')', ')', '/', '(', step.right, '*', 'conj', '(', step.right, ')', ')']),