+++ src/functions/comparison/smaller.js
@@ -35,6 +35,7 @@ class LimeFunctionSmaller extends LimeFunction {
     this.algorithms.set('b(int,int)', (step) => {
       step.bs(this.lime.build('boolean')(step.left.value < step.right.value));
     });
+
     this.algorithms.set('b(rat,rat)', (step) => {
       step.bs(this.lime.direct([step.left.nPlace, '*', step.right.dPlace, '<', step.right.nPlace, '*', step.left.dPlace]));
     });