+++ src/functions/numeric/absolute.js
@@ -20,6 +20,7 @@ class LimeFunctionAbsolute extends LimeFunction {
     this.algorithms.set('r(arg{int[1]})', (step) => {
       step.rus(this.lime.direct([Math.abs(step.right.value[0])]));
     });
+
     this.algorithms.set('r(arg{rat[1]})', (step) => {
       step.rus(this.lime.direct([Math.abs(step.right.value[0].n), '/', Math.abs(step.right.value[0].d)]));
     });