+++ src/functions/numeric/sign.js
@@ -20,6 +20,7 @@ class LimeFunctionSign extends LimeFunction {
     this.algorithms.set('r(arg{int[1]})', (step) => {
       step.rus(this.lime.direct([Math.sign(step.right.value[0])]));
     });
+
     this.algorithms.set('r(arg{rat[1]})', (step) => {
       step.rus(this.lime.direct([Math.sign(step.right.value[0].n / step.right.value[0].d)]));
     });