+++ src/functions/statistical/random.js
@@ -22,9 +22,11 @@ class LimeFunctionRandom extends LimeFunction {
       const m = 10 ** 15;
       step.rus(this.lime.direct([Math.floor(Math.random() * m), '/', m]));
     });
+
     this.algorithms.set('r(arg{int[1]})', (step) => {
       step.rus(this.lime.direct(['rand', '(', 0, ',', step.right.places[0], ')']));
     });
+
     this.algorithms.set('r(arg{int[2]})', (step) => {
       const [l, r] = step.right.places;
       if (this.lime.direct([l.value, '>', r.value]).value) {