+++ src/functions/structural/omitted.js
@@ -20,6 +20,7 @@ class LimeFunctionOmitted extends LimeFunction {
     this.algorithms.set('b(mat,arg{int[1]})', (step) => {
       step.bs(step.lime.direct(['index', '(', step.left, ',', step.right, ')']));
     });
+
     this.algorithms.set('b(expr,expr)', (step) => {
       step.fs(step.lime.refer('*'));
     });